requirements.txt
pandas
psycopg2-binary
sqlalchemy
pytest
python-dotenv
